fix the bug :)
the problem is that you are checking the value of process.domain while not inside domain.enter(). I guess in this case there are no guarantees that it will not have garbage from old sessions. Of course all of the domain related code can be deleted once experiment is merged.